+++ src/prepfold_multi_cmd.c
@@ -68,7 +68,7 @@ static Cmdline cmd = {
     /* absphaseP = */ 0,
     /***** -barypolycos: Force the use of polycos for barycentered events */
     /* barypolycosP = */ 0,
-    /***** -debug: Show debugging output when calling TEMPO for polycos */
+    /***** -debug: Show debugging output when calling tempo2 for polycos */
     /* debugP = */ 0,
     /***** -samples: Treat the data as samples and not as finite-duration integrated data */
     /* samplesP = */ 0,
@@ -670,7 +670,7 @@ void showOptionValues(void)
         printf("-barypolycos found:\n");
     }
 
-    /***** -debug: Show debugging output when calling TEMPO for polycos */
+    /***** -debug: Show debugging output when calling tempo2 for polycos */
     if (!cmd.debugP) {
         printf("-debug not found.\n");
     } else {